# postman/specs/Legal API/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Legal API",
    "version": "1.0.0",
    "description": "API documentation for Legal backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "string",
        "example": "681d0f8f8d7c4b0012f34abc"
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 10
          }
        }
      },
...
      "CaseStageStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "delayed",
          "cancelled"
        ],
        "example": "pending"
      },
...
      "CaseStage": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "caseId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "title": {
            "type": "string",
            "example": "Revisión de documentos"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "assignedTo": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "assignedBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CaseStageStatus"
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "example": "high"
          },
          "estimatedDays": {
            "type": "number",
            "nullable": true,
            "example": 5
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "delayReason": {
            "type": "string",
            "nullable": true
          },
          "dependsOn": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "isFinalStage": {
            "type": "boolean",
            "example": false
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCaseStageDTO": {
        "type": "object",
        "required": [
          "caseId",
          "title"
        ],
        "properties": {
          "caseId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Revisión de documentos"
          },
          "description": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "example": "high"
          },
          "estimatedDays": {
            "type": "number",
            "example": 5
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "dependsOn": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isFinalStage": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "UpdateCaseStageDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "estimatedDays": {
            "type": "number"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "dependsOn": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isFinalStage": {
            "type": "boolean"
          }
        }
      },
      "UpdateCaseStageStatusDTO": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CaseStageStatus"
          },
          "delayReason": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AssignCaseStageDTO": {
        "type": "object",
        "required": [
          "assignedTo"
        ],
        "properties": {
          "assignedTo": {
            "type": "string",
            "example": "681d0f8f8d7c4b0012f34abc"
          }
        }
      },
      "ReorderCaseStageDTO": {
        "type": "object",
        "required": [
          "order"
        ],
        "properties": {
          "order": {
            "type": "number",
            "example": 2
          }
        }
      },
      "CaseStageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "stage": {
            "$ref": "#/components/schemas/CaseStage"
          }
        }
      },
      "CaseStageListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseStage"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "StageStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "example": "Estado de etapa actualizado"
          },
          "stage": {
            "$ref": "#/components/schemas/CaseStage"
          }
        }
      },
      "AssignStageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "example": "Etapa asignada correctamente"
          },
          "stage": {
            "$ref": "#/components/schemas/CaseStage"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/cases/{caseId}/stages": {
      "post": {
        "summary": "Crear una etapa para un caso",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseStageDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Etapa creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "get": {
        "summary": "Obtener etapas de un caso",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de etapas obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStageListResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      }
    },
    "/case-stages/{id}": {
      "get": {
        "summary": "Obtener etapa por ID",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la etapa"
          }
        ],
        "responses": {
          "200": {
            "description": "Etapa obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Etapa no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar etapa",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la etapa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseStageDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Etapa actualizada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseStageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Etapa no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar etapa",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la etapa"
          }
        ],
        "responses": {
          "200": {
            "description": "Etapa eliminada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Etapa eliminada correctamente"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Etapa no encontrada"
          }
        }
      }
    },
    "/case-stages/{id}/status": {
      "patch": {
        "summary": "Actualizar estado de etapa",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la etapa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseStageStatusDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Etapa no encontrada"
          }
        }
      }
    },
    "/case-stages/{id}/assign": {
      "patch": {
        "summary": "Asignar etapa a un usuario",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la etapa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCaseStageDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Etapa asignada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignStageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Etapa no encontrada"
          }
        }
      }
    },
    "/case-stages/{id}/reorder": {
      "patch": {
        "summary": "Reordenar etapa",
        "tags": [
          "Case Stages"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la etapa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderCaseStageDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Etapa reordenada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Etapa reordenada correctamente"
                    },
                    "stage": {
                      "$ref": "#/components/schemas/CaseStage"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Etapa no encontrada"
          }
        }
      }
    }
  },
...
}